switch out namespace (remove old)
