+++ packages/textfield/package.json
@@ -20,7 +20,7 @@
   "dependencies": {
     "@emotion/core": "^10.0.28",
     "@saruni-ui/form": "^0.0.1",
-    "@saruni-ui/theme-next": "^0.0.3",
+    "@saruni-ui/theme": "^0.0.3",
     "react-hook-form": "^5.7.2"
   },
   "devDependencies": {
