--- packages/theme-next/.babelrc
@@ -1,3 +1,0 @@
-{
-  "extends": "../../babel.config.js"
-}
