--- packages/theme-next/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@saruni-ui/theme-next",
-  "version": "0.0.3",
-  "author": "Tambium",
-  "license": "Apache-2.0",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  },
-  "scripts": {
-    "build": "babel src --out-dir dist --extensions '.ts,.tsx'",
-    "check:tsc": "tsc",
-    "delete:build": "rm -rf dist && rm -f -- tsconfig.tsbuildinfo"
-  },
-  "dependencies": {
-    "@emotion/core": "^10.0.27"
-  },
-  "devDependencies": {
-    "react-dom": "^16.12.0"
-  },
-  "peerDependencies": {
-    "react": "^16.12.0"
-  }
-}
