--- packages/theme-next/src/constants.ts
@@ -1,11 +1,0 @@
-export const DEFAULT_THEME_MODE = 'light';
-
-export const font = {
-  family: {
-    ui:
-      'system-ui, -apple-system, Segoe UI, Roboto, Ubuntu, Cantarell, Noto Sans, sans-serif',
-  },
-  size: {
-    body: 14,
-  },
-};
