--- packages/theme-next/src/index.ts
@@ -1,10 +1,0 @@
-import {
-  GlobalThemeProvider,
-  useGlobalTheme,
-} from './components/GlobalThemeProvider';
-import { Reset } from './components/Reset';
-import { createTheme } from './utils/createTheme';
-
-export { createTheme, GlobalThemeProvider, Reset, useGlobalTheme };
-
-export * from './types';
