--- packages/theme-next/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.settings.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src"
-  }
-}
