--- packages/theme/CHANGELOG.md
@@ -1,8 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## 0.0.3 (2020-06-05)
-
-**Note:** Version bump only for package @saruni-ui/theme
