--- packages/theme/src/components/Theme.ts
@@ -1,7 +1,0 @@
-import { createTheme } from '../utils/createTheme';
-import { GlobalThemeTokens } from '../types';
-import { DEFAULT_THEME_MODE } from '../constants';
-
-export default createTheme<GlobalThemeTokens, any>(() => ({
-  mode: DEFAULT_THEME_MODE,
-}));
