+++ packages/theme/src/constants.ts
@@ -1,7 +1,11 @@
 export const DEFAULT_THEME_MODE = 'light';
 
-export const borderRadius = (): number => 4;
-export const gridSize = (): number => 8;
-export const fontSize = (): number => 14;
-export const fontFamily = (): string =>
-  `-apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif`;
+export const font = {
+  family: {
+    ui:
+      'system-ui, -apple-system, Segoe UI, Roboto, Ubuntu, Cantarell, Noto Sans, sans-serif',
+  },
+  size: {
+    body: 14,
+  },
+};
