+++ packages/theme/src/index.ts
@@ -1,10 +1,10 @@
-import * as colors from './colors';
-import * as math from './utils/math';
+import {
+  GlobalThemeProvider,
+  useGlobalTheme,
+} from './components/GlobalThemeProvider';
+import { Reset } from './components/Reset';
+import { createTheme } from './utils/createTheme';
 
-export { colors, math };
+export { createTheme, GlobalThemeProvider, Reset, useGlobalTheme };
 
-export * from './components/Reset';
-export { default } from './components/Theme';
-export * from './constants';
 export * from './types';
-export * from './utils/createTheme';
