+++ packages/theme/src/utils/createTheme.tsx
@@ -1,63 +1,43 @@
-import React, { createContext, ComponentType, ReactNode } from 'react';
-
-export type ThemeProp<ThemeTokens, ThemeProps> = (
-  getTokens: (props: ThemeProps) => ThemeTokens,
-  themeProps: ThemeProps,
-) => ThemeTokens;
+import React from 'react';
+import { ThemeProp } from '../types';
 
 export function createTheme<ThemeTokens, ThemeProps>(
-  defaultGetTokens: (props: ThemeProps) => ThemeTokens,
-): {
-  Consumer: ComponentType<
-    ThemeProps & {
-      children: (tokens: ThemeTokens) => ReactNode;
-    }
-  >;
-  Provider: ComponentType<{
-    children?: ReactNode;
-    value?: ThemeProp<ThemeTokens, ThemeProps>;
-  }>;
-} {
-  const emptyThemeFn: ThemeProp<ThemeTokens, ThemeProps> = (getTokens, props) =>
-    getTokens(props);
+  createTokenFn: (props: ThemeProps) => ThemeTokens,
+) {
+  const emptyThemeFn: ThemeProp<ThemeTokens, ThemeProps> = (baseTokens) =>
+    baseTokens;
 
-  const ThemeContext = createContext(defaultGetTokens);
+  const ThemeContext = React.createContext({
+    createTokenFn,
+    customTheme: emptyThemeFn,
+  });
 
-  function Consumer(
-    props: ThemeProps & { children: (tokens: ThemeTokens) => ReactNode },
-  ) {
-    const { children, ...themeProps } = props;
+  const Provider = (props: {
+    children: React.ReactNode;
+    theme?: ThemeProp<ThemeTokens, ThemeProps>;
+  }) => {
+    const customTheme = props.theme || emptyThemeFn;
+
     return (
-      <ThemeContext.Consumer>
-        {(theme) => {
-          const themeFn = theme || emptyThemeFn;
-          // @ts-ignore
-          const tokens = themeFn(themeProps);
-          return children(tokens);
-        }}
-      </ThemeContext.Consumer>
+      <ThemeContext.Provider
+        {...props}
+        value={{ createTokenFn, customTheme }}
+      />
     );
+  };
+
+  function useTheme(props: ThemeProps) {
+    const { createTokenFn, customTheme } = React.useContext(ThemeContext);
+
+    const baseTokens = createTokenFn(props);
+
+    const customizedTokens = {
+      ...baseTokens,
+      ...customTheme(baseTokens, props),
+    };
+
+    return { tokens: customizedTokens };
   }
 
-  function Provider(props: {
-    children?: ReactNode;
-    value?: ThemeProp<ThemeTokens, ThemeProps>;
-  }) {
-    return (
-      <ThemeContext.Consumer>
-        {(themeFn) => {
-          const valueFn = props.value || emptyThemeFn;
-          const mixedFn = (themeProps: ThemeProps) =>
-            valueFn(themeFn, themeProps);
-          return (
-            <ThemeContext.Provider value={mixedFn}>
-              {props.children}
-            </ThemeContext.Provider>
-          );
-        }}
-      </ThemeContext.Consumer>
-    );
-  }
-
-  return { Consumer, Provider };
+  return { Provider, useTheme };
 }
