--- packages/theme/src/utils/math.ts
@@ -1,15 +1,0 @@
-export function add(fn: (props?: any) => number, addend: number) {
-  return (props?: any) => fn(props) + addend;
-}
-
-export function subtract(fn: (props?: any) => number, subtrahend: number) {
-  return (props?: any) => fn(props) - subtrahend;
-}
-
-export function multiply(fn: (props?: any) => number, factor: number) {
-  return (props?: any) => fn(props) * factor;
-}
-
-export function divide(fn: (props?: any) => number, divisor: number) {
-  return (props?: any) => fn(props) / divisor;
-}
