+++ tsconfig.json
@@ -4,7 +4,6 @@
   "references": [
     { "path": "./packages/form" },
     { "path": "./packages/textfield" },
-    { "path": "./packages/theme" },
-    { "path": "./packages/theme-next" }
+    { "path": "./packages/theme" }
   ]
 }
